Implement keyboard capture
This addresses issue #19. However, main.ts is getting quite unweildy
with spaghetti state management. In addition, there is a conceptual
issue becoming clear between game-rendered input prompts and DOM-level
input prompts. E.g. in the game-over state, the game renders the message
"Space to restart," but now that only works if the game is not "paused,"
aka input is being captured. As another example, the hotkeys identified
by the game-mode checkbox labels only work when input is captured, but
that's not clear from looking at them.

The plan for the state management issue is to rewrite the UI in React,
which will be the next bit of work.

I'm not sure OTTOMH what to do about the conceptual issue, but I think
clarifying state management may help in thinking it through.